refactor: Move NetworkUpgrade to zebra-chain
