+++ zebra-chain/src/lib.rs
@@ -22,6 +22,7 @@ pub mod keys;
 pub mod note_commitment_tree;
 pub mod notes;
 pub mod nullifier;
+pub mod parameters;
 pub mod proofs;
 pub mod serialization;
 pub mod transaction;
@@ -30,6 +31,8 @@ pub mod utils;
 
 pub use ed25519_zebra;
 pub use redjubjub;
+
+pub use parameters::NetworkUpgrade;
 
 #[cfg(test)]
 use proptest_derive::Arbitrary;
